Switch to QueryTree API
In preparation for adding the MultipassPresearcher, the API for extracting terms
form queries is changed.  Instead of examining the query in-place and extracting
terms recursively, we now build a nested QueryTree structure from the query, and
then extract terms from it in a second step.

Decoupling the analysis and extraction steps will allow multiple passes through
the query tree.