Move to PresearcherComponent architecture
A PresearcherComponent is a way of grouping the analysis and retrieval of a set
of queries.  Previously this was achieved by subclassing Presearcher itself, but
moving from inheritance to composition allows us to decouple the types of
queries accepted by a Presearcher, and the sort of analysis we want to perform.

This step is necessary for the MultipassPresearcher.